Create number_formats.py
given an integer. n print the following value for each integer i from 1 to n: 
1. Decimal
2. Octal
3. Hexadecimal (capitalized)
4. Binary

The four values must be printed on a single line in the order specified above for each i from 1 to number and the value should be spaced-padded to match the width of the binary value of number and the values should be separated by a single space.