Create defaultdict.py
In this challenge, you will be given 2 integers, n and m. There are n words, which might repeat in word grout A. There are m words belonging to word group B. For each m words, check whether the word has appeared in group A or not. Print the indices of each occurence of m in grout A. If it does not appear, print -1.

STDIN   Function
 -----     --------
5 2     group A size n = 5, group B size m = 2
a       group A contains 'a', 'a', 'b', 'a', 'b'
a
b
a
b
a       group B contains 'a', 'b'
b

Sample Output

1 2 4
3 5