Create exceptions.py
You are given two values a and b. Perform integer division and print a/b. Use error handling and handle ZeroDivisionError and ValueError.

Input Format:
The first line contains T, the numver of test xases.
The next T lines each contain the space separated values of a and b.

Output:
Print the value of a/b.
In the case of ZeroDIvisionError or ValueError, print the error code.