Create collections_ordereddict.py
You are the manager of a supermarker. 
You have a list of N items together with their prices that consumers bought on a particular day.
YOur task is to print each item_name and net_price in order of its first occurrence. 
item_name - Name of the item.
net_price = Quantity of the item sold multiplieed by the price of each item.

INPUT FORMAT:

The first line contains the number of items, N.
The next N line contains the item's name and price, separated by a space.

OUTPUT: 

Print the item_name and net_price in order of its first occurrence.