Create company_name.py
Task:

A newly opened multinational brand has decided to base their company logo on the three most common characters in the company name. They are now trying out various combinations of company names and logos based on this condition. Given a string s, which is the company name in lowercase letters, your task is to ffind the top three most common characters in the string.

1. Print the three most common characters along wiht thier occurrence count.
2. Sort in descending order of occurrrence count.
3. It the occurrence count is the same, sort the characters in alphabaetical order.

Input Format:

A single line of input containing the string S.

Output Format:

Print the three most common characters along with their occurrence count each on a separate line. Sort output in descending order of occurrence count.
If the occurrence count is the same, sort the characters in alphabetical order.