Create check_strict_SuperSet.py
You are given a set A and n other sets.
Your job is to find whether set A is a strict superset of each of the N sets.
Print True, if A is a strict superset of each of the N sets. Otherwise, print False.
A strict superset has at least one element that doesnot exist in its subset.

Example: Set([1,2,3]) is a strict superset of set([1,2])

Input Format:
The first line contains the space separated elements of set A.
The secnd line cntains integer n, the number of elemtns of the other sets.
The next n lines contains the space separated elements of the other sets.

Outpur Format:
Print True if set A is strice superset of all other N sets. Otherwise, print False.