Create athlets_sort.py
You are given a spreadsheet that contains a list of N athletes and their details (such as age, height, weght and so on). You are required to sort the data based on the Kth attribute and print the final resulting table.
Note that K is indexed from 0 to M - 1, where M is the number of attributes.
Note: If two attributes are the same for different rows, for example, if two atheletes are of the same age, print the row that appeared first in the input.

Input Format:
The first line contains N and M separated by a space.
The next N lines each contain M elements.
The last line contains K.

Output Format:
Print the N lines of sorted tabel. Each line should contain the space separated elements.